refactor: inner 클래스 별도 클래스로 변경
